Renaming lsp service to cpp lsp service
