+++ plugins/cpp_lsp/service/include/cpplspservice/lsp_types.h
