+++ plugins/cpp_lsp/service/src/lsp_types.cpp
@@ -1,4 +1,4 @@
-#include <lspservice/lsp_types.h>
+#include <cpplspservice/lsp_types.h>
 
 namespace cc
 {
