+++ service/CMakeLists.txt
@@ -3,3 +3,4 @@ add_subdirectory(language)
 add_subdirectory(plugin)
 add_subdirectory(project)
 add_subdirectory(workspace)
+add_subdirectory(lsp)
